# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.AppliedPricingRule:
    properties:
      adjust_kind:
        type: string
      adjust_value:
        type: integer
      name:
        type: string
      rule_id:
        type: string
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.BasePriceRequest:
    properties:
      prices:
        additionalProperties:
          format: int64
          type: integer
        example:
          standard: 80000
          vip: 120000
        type: object
    required:
    - prices
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.BasePriceResponse:
    properties:
      price:
        type: integer
      seat_type:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreatePricingRuleRequest:
    properties:
      active:
        description: Omitted defaults to TRUE.
        type: boolean
      adjust_kind:
        enum:
        - percent
        - fixed
        example: percent
        type: string
      adjust_value:
        description: |-
          Signed: negative is a discount, positive a surcharge. No sign is forced —
          that is a business decision left open (PLAN_CAMPAIGN.md section 11).
        example: 10
        type: integer
      day_of_week:
        example: 0
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        example: "23:00"
        type: string
      name:
        example: Weekend surcharge
        maxLength: 255
        minLength: 1
        type: string
      priority:
        minimum: 0
        type: integer
      specific_date:
        example: "2026-01-01"
        type: string
      start_time:
        example: "18:00"
        type: string
    required:
    - adjust_kind
    - name
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PricingQuoteResponse:
    properties:
      applied:
        items:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.AppliedPricingRule'
        type: array
      base:
        type: integer
      final:
        type: integer
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PricingRuleResponse:
    properties:
      active:
        type: boolean
      adjust_kind:
        type: string
      adjust_value:
        type: integer
      created_at:
        type: string
      day_of_week:
        type: integer
      end_time:
        type: string
      id:
        type: string
      name:
        type: string
      priority:
        type: integer
      specific_date:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PublicPriceListResponse:
    properties:
      from_price:
        type: integer
      prices:
        additionalProperties:
          format: int64
          type: integer
        type: object
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdatePricingRuleRequest:
    properties:
      active:
        type: boolean
      adjust_kind:
        enum:
        - percent
        - fixed
        type: string
      adjust_value:
        type: integer
      day_of_week:
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      priority:
        minimum: 0
        type: integer
      specific_date:
        type: string
      start_time:
        type: string
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: success
        type: string
      reason:
        description: |-
          Reason is a stable machine-readable key for an error ("discount_expired"), so the
          frontend can translate without matching on the English sentence. Absent on success.
        type: string
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Meta:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 42
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged:
    properties:
      items: {}
      meta:
        $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Meta'
    type: object
...
info:
  contact: {}
  description: API for the Cinema Project box-office system.
  title: BackEnd-CP API
  version: "1.0"
paths:
...
  /admin/pricing/base:
    get:
      description: One row per seat type. All four seat types always exist once migration
        000013 has run.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.BasePriceResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Get the global base prices (admin)
      tags:
      - pricing
    put:
      consumes:
      - application/json
      description: Send just the seat type(s) you are changing; the rest are left
        alone.
      parameters:
      - description: Seat type -> price
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.BasePriceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.BasePriceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Set the global base price for one or more seat types (admin)
      tags:
      - pricing
  /admin/pricing/rules:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Match rule name
        in: query
        name: search
        type: string
      - description: Filter enabled/disabled; omit for both
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PricingRuleResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List pricing rules (admin)
      tags:
      - pricing
    post:
      consumes:
      - application/json
      parameters:
      - description: Rule
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreatePricingRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PricingRuleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Create a pricing rule (admin)
      tags:
      - pricing
  /admin/pricing/rules/{id}:
    delete:
      parameters:
      - description: Pricing rule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Delete a pricing rule (admin)
      tags:
      - pricing
    get:
      parameters:
      - description: Pricing rule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PricingRuleResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Get one pricing rule (admin)
      tags:
      - pricing
    patch:
      consumes:
      - application/json
      description: PARTIAL update. Sending "" for start_time/end_time/specific_date
        clears it back to NULL.
      parameters:
      - description: Pricing rule id
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdatePricingRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PricingRuleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Update a pricing rule (admin)
      tags:
      - pricing
...
  /pricing:
    get:
      description: 'The global base price for each seat type, plus the cheapest configured
        seat anywhere. No auth: this is the customer price page. Replaces the old
        per-hall price list (PLAN_CAMPAIGN.md section 11.4, Phase 3) now that hall_prices
        is gone - every hall shares the same prices.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PublicPriceListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      summary: Public ticket price list
      tags:
      - pricing
  /pricing/quote:
    get:
      description: 'Read-only: base price, every rule that matched (in applied order)
        and the floored-at-0 final price. Never touches booking or hold state.'
      parameters:
      - description: Showtime id
        in: query
        name: showtime_id
        required: true
        type: string
      - description: standard, vip, couple or recliner
        in: query
        name: seat_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.PricingQuoteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: showtime not found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      summary: Preview the price for one showtime and seat type
      tags:
      - pricing
...
securityDefinitions:
  BearerAuth:
    description: 'Header format: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"